\name{densityplot_all}
\alias{densityplot_all}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Density plots for microarrays
}
\description{
This function makes density plots from the object ds generated by aroma.affymetrix
}
\usage{
densityplot_all(ds, labels, picname, groupn = NULL, all = FALSE, resDir = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ds}{
Object obtained with aroma.affymetrix and the function AffymetrixCelSet in the aroma.affymetrix package.
}
  \item{labels}{
Optional. Vector with sample names. Samples are in the same order as in the ds object. Default get sample names from the ds object.
}
  \item{picname}{
Name of output files.
}
  \item{groupn}{
Optional. Number of samples to include in each density plot. If the number is lower than total number of samples, more than one plot is going to be shown. Default is groupn = all samples.
}
  \item{all}{
Optional. If TRUE, an additional plot with all density plots made with a groupn is made.
}
  \item{resDir}{
Output directory. Default is resultsDir.
}
}
\details{
This function outputs density plots from a ds object obtained with aroma.affymetrix and the function AffymetrixCelSet. Plots are made using the log2 transformation of the intensity signal. If a large amount of samples is contained in the ds object, it could be dificult to distinguish one from another. Using the additional parameter groupn, a lower number of samples can be defined to be ploted. For example if there are 24 samples in the ds object, seting groupn=12, two plots of 12 samples each will be made. If all=TRUE, all plots generated by the function will be shown in the same plot.
}
\value{
Output is a .png file with the index of the arrays in the title in resDir.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Magdalena Arnal Segura <marnal@imim.es>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}

