% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pheat.make.function.R
\name{PHEAT.MAKE}
\alias{PHEAT.MAKE}
\title{Function to make heatmaps from a numeric matrix using pheatmap package}
\usage{
PHEAT.MAKE(matrix,cols.s = NULL,filename,filedir,rownames.vect = "",cex.col = 0.8,scale = "none",toPNG = FALSE,annotation_col = NULL,annotation_row = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{matrix}{
    Numeric matrix with the intensity values to plot
  }
  \item{cols.s}{
    List for specifying annotation_row and annotation_col track colors manually. It is possible to define the colors for only some of the features
  }
  \item{filename}{
    Output file name
  }
  \item{filedir}{
    Output directory, it can be a pathway
  }
  \item{rownames.vect}{
    Vector with the names of the genes in the heatmap
  }
  \item{cex.col}{
    Size of the column labels
  }
  \item{scale}{
    Character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are "row", "column" and "none"
  }
  \item{toPNG}{
     TRUE if we want .png output, FALSE if we want .pdf output
  }
  \item{annotation_row}{
    Data frame that specifies the annotations shown on left side of the heatmap. Each row defines the features for a specific row
  }
    \item{annotation_col}{
    Similar to annotation_row, but for columns
  }

}
\details{
  This function makes a plot in png format using the results from the differential expression analysis with "limma" package. It uses the columns "adj.P.Val" or "P.Value" deppending on the value "adj" and the column logFC. Cut-off is used in order to indicate the DE genes in the scatterplot.
}
\value{
  The plot is created in the "filePath" directory with the name "fileName" in ".png" format.
}
\author{
  Ariadna Acedo Terrades <aacedo@imim.es>.
}
