\name{QualityGraphs-package}
\alias{QualityGraphs-package}
\alias{QualityGraphs}
\docType{package}
\title{
\packageTitle{QualityGraphs}
}
\description{
\packageDescription{QualityGraphs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QualityGraphs}
\packageIndices{QualityGraphs}

}
\author{
\packageAuthor{QualityGraphs}

Maintainer: \packageMaintainer{QualityGraphs}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
