% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq.metrics.R
\name{read.RNA_Metrics}
\alias{read.RNA_Metrics}
\title{Read log files from Picard CollectRnaSeqMetrics tool}
\usage{
read.RNA_Metrics(rnametricsDir)
}
\arguments{
\item{rnametricsDir}{Character vector with the directory where CollectRnaSeqMetrics 
output files are stored}
}
\value{
Numeric dataframe with one row per file and columns:  percentage of ribosomal,  
coding, UTR, intronic, intergenic and mRNA bases.
}
\description{
Read log files from Picard CollectRnaSeqMetrics tool
}
\author{
Julia Perera Bel <jperera@imim.es>
}
