% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeq.metrics.R
\name{RNASeq.metrics.plots}
\alias{RNASeq.metrics.plots}
\title{Make RNASeq metrics plots for alignment assessment}
\usage{
RNASeq.metrics.plots(
  starDir = NULL,
  rnametricsDir = NULL,
  counts = NULL,
  resDir = getwd(),
  picname = "RNASeq.metrics",
  toPNG = F
)
}
\arguments{
\item{starDir}{Character vector with the directory where STAR 
*Log.final.out files are stored. If more than one run, provide a list with
all directories eg. list("/dir/run1","/dir/run2","/dir/run3").}

\item{rnametricsDir}{Character vector with the directory where 
CollectRnaSeqMetrics output files are stored. Files must be named *.RNA_Metrics.}

\item{counts}{Data frame with annotations (first 7 columns) and counts 
(one column per sample), as read from CountsTable.txt from FeatureCounts. 
Only needed if "Counts by Chromosome" plot is desired.}

\item{resDir}{Character vector with output results directory. Default is working directory.}

\item{picname}{Character vector with the name of output files.}

\item{toPNG}{Generate additional plots in png format  Default FALSE.}
}
\value{
Files are created in the resDir directory with the picname and .png or .pdf extensions.
}
\description{
Function that returns up to three plots for the assessment of 
RNASeq alignment quality. Possible plots are: 1. read mapping stats provided 
by STAR (eg. uniquely, multimapped), 2. mapping by genomic regions (eg. coding, 
intergenic, intronic) provided by Picard tools CollectRNASeqMetrics and 
3. counts by chromosome (X, Y and M are shown).
}
\author{
Julia Perera Bel <jperera@imim.es>
}
