\name{densityplot_all}
\alias{densityplot_all}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Density plots for microarrays
}
\description{
This function makes density plots from the object ds generated by aroma.affymetrix
}
\usage{
densityplot_all(ds, labels, picname, groupn = NULL, all = FALSE, resDir = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ds}{
Object obtained with aroma.affymetrix and the function AffymetrixCelSet in the aroma.affymetrix package
}
  \item{labels}{
Optional. Vector with sample names in the same order as in the ds object
}
  \item{picname}{
Name of the output
}
  \item{groupn}{
Optional. Number of samples to include in each plot. If the number is lower than total number of samples, more than one plot is going to be shown
}
  \item{all}{
Optional. If TRUE, an additional plot with all density plots made with a groupn is made.
}
  \item{resDir}{
Output directory. Default resultsDir
}
}
\details{
This function outputs density plots from a ds object obtained with aroma.affymetrix and the function AffymetrixCelSet. Plots are made using the log2 transformation of the intensity signal. If a large amount of samples is contained in the ds object, it could be dificult to distinguish one from another. Using the additional parameter groupn, a lower number of samples can be defined to be ploted. For example if there are 24 samples in the ds object, seting groupn=12, two plots of 12 samples each will be made. If all=TRUE, all plots generated by the function will be shown in the same plot.
}
\value{
Output is a .png file with the index of the arrays in the title.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Magdalena Arnal <marnal@imim.es>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
