% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePCA.2D.R
\name{makePCA.2D}
\alias{makePCA.2D}
\title{Make 2D PCA from a numeric matrix}
\usage{
makePCA.2D(
  est_noctrls,
  picname,
  conditions = NULL,
  colors = NULL,
  dist = 2,
  resDir = NULL
)
}
\arguments{
\item{est_noctrls}{numeric matrix of intensities or normalized counts (eg. TMM log2CPM)}

\item{picname}{character vector with the name of output files.}

\item{conditions}{vector with the conditions (same length as colnames(est_noctrls))}

\item{colors}{vector with the colors assigned to each condition (in order of the unique(conditions))}

\item{dist}{distance between the dots of the PCa and the labels}

\item{resDir}{character vector with output results directory. Default is working directory.}
}
\value{
files are created in the resDir directory with the picname and .png extension.
}
\description{
Function that returns a 2D PCA plot with PC1 in x axis and PC2 in y axis.
}
\author{
Julia Perera Bel <jperera@imim.es>
}
